% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_le_effects.R
\name{compute_le_effects}
\alias{compute_le_effects}
\title{Compute \dQuote{Least Extreme} Effects}
\usage{
compute_le_effects(effects_matrix)
}
\arguments{
\item{effects_matrix}{The input effects matrix in which rows
correspond to the samples/observations and columns correspond
to thedifferent signals.}
}
\value{
A matrix of the same dimension as the input matrix
  containing the least extreme (l.e.) effects.
}
\description{
This function is useful for identifying
  \dQuote{distinctive} effects; that is, effects that are very
  different (much larger or much smaller) than the effects in the
  other signals.
}
\examples{
X <- rbind(c(4,0,0),
           c(1,-1,0),
           c(4,0,3),
           c(2,0,-4),
           c(0,-4,0),
           c(0,-4,-3),
           c(0,-2,-4),
           c(1,-2,4),
           c(0,0,0))
rownames(X) <- paste0("row",1:9)
colnames(X) <- paste0("col",1:3)
Y <- compute_le_effects(X)

}
