% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_factors_nn_to_pn.R
\name{convert_factors_nn_to_pn}
\alias{convert_factors_nn_to_pn}
\title{Convert Some Non-Negative Factors to Positive/Negative Factors}
\usage{
convert_factors_nn_to_pn(fl, kset, data, ...)
}
\arguments{
\item{fl}{A flashier fit with non-negative factors, such as an
output from a call to \link{flashier_nmf}.}

\item{kset}{The indices of the non-negative factors to convert to
positive/negative factors.}

\item{data}{The data matrix passed as an input to
\code{\link[flashier]{flash_init}}.}

\item{\dots}{Additional arguments passed to
\code{\link[flashier]{flash_init}}.}
}
\value{
A \sQuote{flash} object.
}
\description{
Assuming the input flash model has non-negative
  factors with point-exponential priors, this function creates a
  new flash model in which selected factors have
  point-exponential priors for the columns of L and point-Laplace
  priors for the corresponding columns of F, which allows for the
  parameter estimates to be either positive or negative. Note
  that L and F in the new flash model are initialized to the
  estimates from the input flash model, so initially they are all
  non-negative; a subsequent call to
  \code{\link[flashier]{flash_backfit}} is needed to actually
  produce negative values.
}
