% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_heatmap.R
\name{effects_heatmap}
\alias{effects_heatmap}
\title{Add Title Here}
\usage{
effects_heatmap(
  effects_matrix,
  zero_value = 0.01,
  size_range = c(2, 6),
  font_size = 10
)
}
\arguments{
\item{effects_matrix}{n x d numeric matrix, where n is the number
of features and d is the number of different signals.}

\item{zero_value}{Numbers smaller than \code{zero_value} (in
magnitude) are not shown in the plot.}

\item{size_range}{Passed as the \dQuote{range} argument to
\code{\link[ggplot2]{scale_size}}.}

\item{font_size}{Specifies the font size for the plot.}
}
\value{
Describe the return value here.

A \code{ggplot} object.
}
\description{
This is a generic plotting utility for comparing the
  \dQuote{effects} on different features (rows) within a given
  signal (column), and for comparing the effect of the same
  feature across different signals.
}
\examples{
set.seed(1)
X <- matrix(sample(seq(-5,5),200,replace = TRUE),20,10)
rownames(X) <- paste0("row",1:20)
colnames(X) <- paste0("col",1:10)
effects_heatmap(X,size_range = c(2,8),font_size = 9)
effects_heatmap(abs(X),size_range = c(2,8),font_size = 9)

}
