% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashier_nmf.R
\name{flashier_nmf}
\alias{flashier_nmf}
\title{Title Goes Here.}
\usage{
flashier_nmf(data, k, maxiter = 100, n.threads = 1, verbose = 1, ...)
}
\arguments{
\item{data}{Describe the data input here.}

\item{k}{Describe the k input here.}

\item{maxiter}{Describe the maxiter input here.}

\item{n.threads}{Describe the n.threads input here.}

\item{\dots}{Additional arguments passed to ...}
}
\value{
Describe the return value here.
}
\description{
Add description of the function here.
}
\examples{
# Add a flashier_nmf example.

}
