% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Print Hello World}
\usage{
hello_world(n, verbose = TRUE)
}
\arguments{
\item{n}{The number of times to say \dQuote{Hello World!}}

\item{verbose}{If \code{verbose = TRUE}, print the messages to the
console.}
}
\value{
A character vector containing the message.
}
\description{
Prints a \dQuote{Hello World} message to the console
  and provides the same message in the output.
}
