% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{perform_gsea}
\alias{perform_gsea}
\title{Add Title Here}
\usage{
perform_gsea(gene_signals, gene_sets, min_size = 10, max_size = 400)
}
\arguments{
\item{gene_signals}{Describe the gene_signals input here.}

\item{gene_sets}{}
}
\value{
Describe the return value here.
}
\description{
Add description here.
}
\examples{
# Add an example here illustrating the use of
# perform_gsea().

}
