% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_effects.R
\name{rank_effects}
\alias{rank_effects}
\title{Rank Effects Matrix}
\usage{
rank_effects(effects_matrix)
}
\arguments{
\item{effects_matrix}{The input effects matrix in which rows
correspond to the samples/observations and columns correspond
to thedifferent signals.}
}
\value{
A matrix of the same dimension as the input matrix
  containing the rankings.
}
\description{
Rank the effects in each column of the input matrix,
  from largest (rank of 1) to smallest (rank of n, where n is the
  number of rows).
}
\examples{
X <- matrix(c(1:5,10:6,-(1:5)),5,3)
rownames(X) <- paste0("row",1:5)
colnames(X) <- paste0("col",1:3)
Y <- rank_effects(abs(X))

}
