% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_transform_effects_matrix.R
\name{rank_transform_effects_matrix}
\alias{rank_transform_effects_matrix}
\title{Rank-Transform Effects Matrix}
\usage{
rank_transform_effects_matrix(effects_matrix, compare_cols = FALSE)
}
\arguments{
\item{effects_matrix}{The n x m matrix X.}

\item{compare_cols}{Explain here what \code{compare_cols} = TRUE does.}
}
\value{
The n x m rank-transformed effects matrix.
}
\description{
For n x m effects matrix X, this function transforms
  each column of X so that the values lie between 0 and 1, with 0
  being the lowest rank, and 1 being the highest. By
  "highest rank", we mean that the entry has the largest (most
  positive) value amongst all the entries in the column.
}
