% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_le_effects.R
\name{compute_le_effects}
\alias{compute_le_effects}
\title{Add Title Here.}
\usage{
compute_le_effects(effects_matrix)
}
\arguments{
\item{effects_matrix}{Describe the effects matrix argument here.}
}
\value{
Describe the return value here.
}
\description{
Write description of this function here.
}
\examples{
X <- rbind(c(4,0,0),
           c(1,-1,0),
           c(4,0,3),
           c(2,0,-4),
           c(0,-4,0),
           c(0,-4,-3),
           c(0,-2,-4),
           c(1,-2,4),
           c(0,0,0))
rownames(X) <- paste0("row",1:9)
colnames(X) <- paste0("col",1:3)
Y <- compute_le_effects(X)

}
