% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_factors_nn_to_pn.R
\name{convert_factors_nn_to_pn}
\alias{convert_factors_nn_to_pn}
\title{Title Goes Here}
\usage{
convert_factors_nn_to_pn(fl, kset, dat, ...)
}
\arguments{
\item{fl}{Describe the fl input argument here.}

\item{kset}{Describe the kset input argument here.}

\item{dat}{Describe the dat input argument here.}

\item{\dots}{Additional parameters.}
}
\value{
Describe the return value here.
}
\description{
Description goes here.
}
