% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_factors_nn_to_pn.R
\name{convert_factors_nn_to_pn}
\alias{convert_factors_nn_to_pn}
\title{Convert Some Non-Negative Factors to Positive/Negative Factors}
\usage{
convert_factors_nn_to_pn(fl, kset, data, ...)
}
\arguments{
\item{fl}{A flashier fit with non-negative factors, such as an
output from a call to \link{flashier_nmf}.}

\item{kset}{The indices of the non-negative factors to convert to
positive/negative factors.}

\item{data}{The data matrix passed as an input to
\code{\link[flashier]{flash_init}}.}

\item{\dots}{Additional arguments passed to
\code{\link[flashier]{flash_init}}.}
}
\value{
A \sQuote{flash} object.
}
\description{
Description goes here.
}
