% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashier_nmf.R
\name{flashier_nmf}
\alias{flashier_nmf}
\title{Fit flashier NMF model by Initializing with NNLM}
\usage{
flashier_nmf(data, k, maxiter = 100, n.threads = 1, verbose = 1, ...)
}
\arguments{
\item{data}{The data matrix. It may be a sparse or dense matrix,
however the initialization step will convert the sparse matrix
to a dense matrix (and will issue a warning when doing so).}

\item{k}{The numbe of factors in the matrix factorization.}

\item{maxiter}{The maximum number of backfitting iterations in each
of the two calls to \code{\link[flashier]{flash_backfit}}.}

\item{n.threads}{The number of threads/CPUs used in
\code{\link[NNLM]{nnmf}}.}

\item{verbose}{This is the setting "verbose" for both the
\code{\link[NNLM]{nnmf}} and
\code{\link[flashier]{flash_backfit}} calls.}

\item{\dots}{Additional arguments passed to
\code{\link[flashier]{flash_init}}.}
}
\value{
A \sQuote{flash} object.
}
\description{
Since the \dQuote{greedy} initialization approach
  implemented in flashier does not seem to work well in some
  cases, this function uses a different initialization strategy:
  obtain a \dQuote{good} initialization using the NNLM package,
  then use this initialization to fit an NMF using flashier.
}
\examples{
library(flashier)
set.seed(1)
n <- 100
m <- 200
k <- 3
L <- matrix(runif(n*k),n,k)
F <- matrix(runif(m*k),m,k)
E <- matrix(runif(n*m,0,0.01),n,m)
X <- tcrossprod(L,F) + E
fl <- flashier_nmf(X,k = 3)
L_est <- ldf(fl,type = "i")$L
F_est <- ldf(fl,type = "i")$F
ks <- c(1,3,2)
plot(L,L_est[,ks],pch = 20,xlab = "true",ylab = "estimated")
abline(a = 0,b = 1,lty = "dotted",col = "magenta")
plot(F,F_est[,ks],pch = 20,xlab = "true",ylab = "estimated")
abline(a = 0,b = 1,lty = "dotted",col = "magenta")

}
