% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashier_nmf.R
\name{flashier_nmf}
\alias{flashier_nmf}
\title{Title Goes Here.}
\usage{
flashier_nmf(data, k, maxiter = 100, n.threads = 1, verbose = 1, ...)
}
\arguments{
\item{data}{Describe the data input here.}

\item{k}{Describe the k input here.}

\item{maxiter}{Describe the maxiter input here.}

\item{n.threads}{Describe the n.threads input here.}

\item{verbose}{Describe the verbose input here.}

\item{\dots}{Additional arguments passed to ...}
}
\value{
Describe the return value here.
}
\description{
Add description of the function here.
}
\examples{
library(flashier)
set.seed(1)
n <- 100
m <- 200
k <- 3
L <- matrix(runif(n*k),n,k)
F <- matrix(runif(m*k),m,k)
E <- matrix(runif(n*m,0,0.01),n,m)
X <- tcrossprod(L,F) + E
fl <- flashier_nmf(X,k = 3)
L_est <- ldf(fl,type = "i")$L
F_est <- ldf(fl,type = "i")$F
ks <- c(1,3,2)
plot(L,L_est[,ks],pch = 20,xlab = "true",ylab = "estimated")
abline(a = 0,b = 1,lty = "dotted",col = "magenta")
plot(F,F_est[,ks],pch = 20,xlab = "true",ylab = "estimated")
abline(a = 0,b = 1,lty = "dotted",col = "magenta")

}
