% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Add title here.}
\usage{
hello_world(n, verbose = TRUE)
}
\arguments{
\item{n}{Describe this argument here.}

\item{verbose}{Describe this argument here.}
}
\value{
Describe return value here.
}
\description{
Add description here.
}
