% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{perform_gsea}
\alias{perform_gsea}
\title{Add Title Here}
\usage{
perform_gsea(
  gene_signals,
  gene_sets,
  min_size = 10,
  max_size = 400,
  L = 10,
  coverage = 0.95,
  max_iter = 100,
  tol = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{gene_signals}{rows = genes, columns = signals}

\item{gene_sets}{rows = genes, columns = gene sets}

\item{min_size}{Describe the min_size input here.}

\item{max_size}{Describe the max_size input here.}

\item{L}{Describe the L argument here.}

\item{coverage}{Describe the coverage argument here.}

\item{max_iter}{Describe the max_iter argument here.}

\item{verbose}{Describe the verbose argument here.}

\item{Describe}{hte tol argument here.}
}
\value{
Describe the return value here.
}
\description{
Add description here.
}
\examples{

# Add an example here illustrating the use of perform_gsea().

}
