% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{perform_gsea}
\alias{perform_gsea}
\title{Perform Gene-set Enrichment Analyses Using SuSiE}
\usage{
perform_gsea(
  gene_signals,
  gene_sets,
  gene_set_info = NULL,
  min_size = 10,
  max_size = 400,
  top_genes = 10,
  L = 10,
  coverage = 0.95,
  max_iter = 100,
  tol = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{gene_signals}{A matrix of signals in which rows correspond to
genes and columns correspond to the different signals. If there
is only one signal, this can be a vector.}

\item{gene_sets}{A binary matrix (sparse or dense) of gene sets in
which rows correspond to genes and columns correspond to gene
sets. An entry of 1 means the gene is a member of the gene set.}

\item{gene_set_info}{Optional data frame containing additional
information about the gene sets.}

\item{min_size}{Gene sets with fewer genes than this will not be
considered.}

\item{max_size}{Gene sets with more genes than this will not be
considered.}

\item{top_genes}{The number of \dQuote{top genes} to include in the}

\item{L}{The maximum number of selected gene sets; passed as the
\dQuote{L} argument to \code{\link[susieR]{susie}}.}

\item{coverage}{The \dQuote{coverage} parameter in
\code{\link[susieR]{susie}}.}

\item{max_iter}{The \dQuote{max_iter} parameter in
\code{\link[susieR]{susie}}.}

\item{tol}{The \dQuote{tol} parameter in
\code{\link[susieR]{susie}}.}

\item{verbose}{If \code{verbose = TRUE}, print updates about 
progress of the analysis.}
}
\value{
A list containing (a) the SuSiE model fits (one for each
  signal); (b) a tibble containing the gene set enrichment results.
}
\description{
Add description here.
}
\examples{

# Add an example here illustrating the use of perform_gsea().
library(pathways)
set.seed(1)
data(gene_sets_human)
i <- which(!is.na(gene_sets_human$gene_info$Ensembl))
y <- gene_sets_human$gene_sets[i,"M973"] + 0.1 * rnorm(length(i))
out <- perform_gsea(y,gene_sets_human$gene_sets[i,])
out$selected_gene_sets

}
